Update FirestoreHandler to separate CIR logic into its own folder, add a close connection method and update app.py to call the close connection method
