+++ eq_cims_management_ui/utils/database/firestore_handler.py
@@ -6,22 +6,20 @@ Classes:
 
 Raises:
     RetryError
-    requests.exceptions.ConnectionError
-    ValueError
 """
 
 import logging
-import os
 import uuid
 from datetime import datetime
 from zoneinfo import ZoneInfo
 
-import requests
 from google.api_core.exceptions import RetryError
 from google.api_core.retry import Retry
 from google.cloud.firestore import Client
 from google.cloud.firestore_v1 import Query
 from google.cloud.firestore_v1.base_document import BaseDocumentReference
+
+from eq_cims_management_ui.utils.database.cir_operations import check_cir_status, get_ci_metadata
 
 logger = logging.getLogger(__name__)
 
@@ -34,8 +32,6 @@ class FirestoreHandler:
         create_database_session
         retrieve_latest_session
         set_document_reference
-        check_cir_status
-        get_ci_metadata
     """
 
     def __init__(self) -> None:
@@ -51,8 +47,8 @@ class FirestoreHandler:
         session_id = str(uuid.uuid4())
         latest_session_document_ref = self.client.collection("sessions").document(session_id)
 
-        self.check_cir_status()
-        ci_metadata = self.get_ci_metadata()
+        check_cir_status()
+        ci_metadata = get_ci_metadata()
 
         try:
             logger.info("Creating session in Firestore database...")
@@ -123,47 +119,9 @@ class FirestoreHandler:
         """
         self.latest_session_document_ref = document_reference
 
-    @staticmethod
-    def check_cir_status() -> None:
+
+    def close_connection(self) -> None:
         """
-        Checks the availability of CIR by making a GET request to the '/status' endpoint of CIR API. If the request
-        returns a non-200 status code, an exception is raised.
-
-        Raises:
-            requests.exceptions.ConnectionError: If the request returns a non-200 status code.
+        Closes the connection to the Firestore database by deleting the client instance.
         """
-        try:
-            logger.info("Checking CIR status endpoint...")
-            cir_status_url = f"http://{os.getenv("CIR_API_BASE_URL")}/status"
-            status = requests.get(cir_status_url, timeout=15)
-            if status.status_code == 200:
-                logger.info("Successfully checked CIR status endpoint.")
-
-        except requests.exceptions.ConnectionError as error:
-            logger.exception("Failed to connect to CIR.")
-            raise requests.exceptions.ConnectionError from error
-
-    @staticmethod
-    def get_ci_metadata() -> list[dict]:
-        """
-        Retrieves collection instrument metadata from CIR by making a GET request to the CIR API. The metadata is
-        returned as a list of collection instrument metadata objects. If the response from CIR is empty, an exception
-        is raised.
-
-        Returns:
-            ci_metadata: A list of collection instrument metadata to be added to the Firestore database.
-
-        Raises:
-            ValueError: If the response from CIR is empty.
-        """
-        logger.info("Retrieving collection instrument metadata from CIR...")
-        cir_metadata_url = f"http://{os.getenv("CIR_API_BASE_URL")}/collection-instruments/metadata"
-        metadata_received = requests.get(cir_metadata_url, timeout=15)
-        ci_metadata = metadata_received.json()
-
-        # CI metadata is always returned as a list of dictionaries given CIs are published
-        if not isinstance(ci_metadata, list) or ci_metadata[0].get("message") == "No CI found":
-            logger.error("Failed to retrieve collection instrument metadata from CIR.")
-            raise ValueError
-
-        return ci_metadata
+        self.client.close()
